Move wrapped_axis out of maths.py
